Fixed merge from file folder migrate
